+++ tests/test_openapi_security.py
@@ -60,6 +60,39 @@ def test_apikey_auth_security_scheme(app, client):
     }
 
 
+def test_custom_security_scheme_name(app, client):
+    basic_auth = HTTPBasicAuth(security_scheme_name='basic_auth')
+    token_auth = HTTPTokenAuth(header='X-API-Key', security_scheme_name='myToken')
+
+    @app.get('/foo')
+    @app.auth_required(basic_auth)
+    def foo():
+        pass
+
+    @app.get('/bar')
+    @app.auth_required(token_auth)
+    def bar():
+        pass
+
+    rv = client.get('/openapi.json')
+    assert rv.status_code == 200
+    validate_spec(rv.json)
+    assert 'basic_auth' in rv.json['components']['securitySchemes']
+    assert 'myToken' in rv.json['components']['securitySchemes']
+    assert rv.json['components']['securitySchemes']['basic_auth'] == {
+        'type': 'http',
+        'scheme': 'basic'
+    }
+    assert rv.json['components']['securitySchemes']['myToken'] == {
+        'type': 'apiKey',
+        'name': 'X-API-Key',
+        'in': 'header'
+    }
+    print(rv.json)
+    assert 'basic_auth' in rv.json['paths']['/foo']['get']['security'][0]
+    assert 'myToken' in rv.json['paths']['/bar']['get']['security'][0]
+
+
 def test_unknown_auth_security_scheme(app):
     from flask_httpauth import HTTPDigestAuth
     auth = HTTPDigestAuth()